:art: Update type references
